refactor: App, Title 컴포넌트를 분리했습니다
