+++ src/components/Title.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function Title() {
+  return (
+    <h3>Counter</h3>
+  );
+}
